requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
line-bot-sdk==1.5.0
MarkupSafe==1.0
psycopg2==2.7.4
python-dotenv==0.7.1
requests==2.18.4
SQLAlchemy==1.2.2
urllib3==1.22
Werkzeug==0.14.1
PyJWT==1.5.3